requirements.txt
# FastAPI – веб-фреймворк для створення API
fastapi==0.115.12

# Uvicorn – ASGI-сервер для запуску FastAPI
uvicorn[standard]==0.34.1

# OpenAI – бібліотека для роботи з OpenAI API (наприклад, для генерації тексту)
openai==1.77.0

# python-dotenv – дозволяє завантажувати налаштування з файлу .env (наприклад, ключі API)
python-dotenv==1.1.0

# з Firebase через Python
firebase-admin==6.7.0
google-cloud-firestore==2.20.2
google==3.0.0

PyJWT==2.10.1
passlib==1.7.4
python-jose==3.4.0

# uuid
# firebase
pydantic==2.11.3

bcrypt==3.2.2






